requirements.txt

flask
onnxruntime
Pillow
numpy
torchvision 
